deployment/sandbox-v2/utils/db_migration/requirements.txt
psycopg2-binary==2.8.6
python-dotenv==0.12.0
